+++ .gitignore
@@ -1,0 +1,3 @@
+backend/node_modules/
+frontend/node_modules/
+.env
