+++ frontend/src/components/ProgressChart.jsx
@@ -1,27 +1,27 @@
-import React from 'react';
-import { Bar } from 'react-chartjs-2';
-import { Chart as ChartJS, CategoryScale, LinearScale, BarElement, Title, Tooltip, Legend } from 'chart.js';
+// import React from 'react';
+// import { Bar } from 'react-chartjs-2';
+// import { Chart as ChartJS, CategoryScale, LinearScale, BarElement, Title, Tooltip, Legend } from 'chart.js';
 
-ChartJS.register(
-  CategoryScale,
-  LinearScale,
-  BarElement,
-  Title,
-  Tooltip,
-  Legend
-);
+// ChartJS.register(
+//   CategoryScale,
+//   LinearScale,
+//   BarElement,
+//   Title,
+//   Tooltip,
+//   Legend
+// );
 
 
-const ProgressChart = ({ data }) => {
-  const chartData = {
-    labels: data.map(item => item.date),
-    datasets: [{
-      label: 'Study Hours',
-      data: data.map(item => item.hours),
-      backgroundColor: 'rgba(177, 5, 5, 0.79)',
-    }],
-  };
+// const ProgressChart = ({ data }) => {
+//   const chartData = {
+//     labels: data.map(item => item.date),
+//     datasets: [{
+//       label: 'Study Hours',
+//       data: data.map(item => item.hours),
+//       backgroundColor: 'rgba(177, 5, 5, 0.79)',
+//     }],
+//   };
 
-  return <Bar data={chartData} />;
-};
-export default ProgressChart;
+//   return <Bar data={chartData} />;
+// };
+// export default ProgressChart;
