+++ frontend/src/main.jsx
@@ -2,7 +2,7 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import './index.css'
 import App from './App.jsx'
-import Dashboard from './pages/Dashboard.jsx'
+// import Dashboard from './pages/Dashboard.jsx'
 createRoot(document.getElementById('root')).render(
   <StrictMode>
     <App />
